Create pl_PL.json
Polish localization. 100% complete, unless I missed a line somewhere. 
Numbers (for example in lines 21-27) entered in the following format 1/x2-x4 (so 2-4 12-24, 22-24 and so on)/0 and x5-x9 due to the nature of Polish grammar. 
Consulted with Dan beforehand.
Also, where are the months used? (lines 206-217) Depending on the  context another case may need to be used.